Fix the ball glitch
Changing the sign of the inequality will fix the problem of the glitch. Originally, the "<=" at line 124 would result in an addscore() if the ball is below the enemy paddle. However, we need the addscore() to be called when the ball is above the enemy paddle, changing the sign to ">=" will fix the ball glitch.